pyproject.toml
[project]
name = "scripts.ghstars-ng"
version = "0.1.0"
description = "Shared workspace service for arXiv sync, GitHub linking, enrichment, and export"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.5",
    "aiohttp>=3.13.3",
    "fastapi>=0.116.0",
    "psycopg[binary]>=3.2.9",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.43",
    "uvicorn[standard]>=0.35.0",
]

[dependency-groups]
dev = [
    "anyio>=4.12.0",
    "httpx>=0.28.1",
    "pytest>=8.4.1",
]

[tool.uv]
package = false

[[tool.uv.index]]
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple"
default = true

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests_v2"]
